+++ src/renderer/states/State.ts
@@ -1,0 +1,6 @@
+export class State {
+  clipboard: string = ""
+  translated: string = ""
+  waiting: boolean = false
+  enabled: boolean = true
+}
